+++ saleor/graphql/product/filters.py
@@ -110,9 +110,44 @@ def _clean_product_attributes_range_filter_input(filter_value, queries):
         queries[attr_pk] += attr_val_pks
 
 
-def _clean_product_attributes_boolean_filter_input(
-    filter_value, queries: T_PRODUCT_FILTER_QUERIES
+def _clean_product_attributes_date_time_range_filter_input(
+    filter_value, queries, *, is_date=False
 ):
+    attribute_slugs = [slug for slug, _ in filter_value]
+    attributes = Attribute.objects.filter(slug__in=attribute_slugs).prefetch_related(
+        "values"
+    )
+    values_map = {
+        attr.slug: {
+            "pk": attr.pk,
+            "values": {val.date_time: val.pk for val in attr.values.all()},
+        }
+        for attr in attributes
+    }
+    for attr_slug, val_range in filter_value:
+        attr_pk = values_map[attr_slug]["pk"]
+        attr_values = values_map[attr_slug]["values"]
+        gte = val_range.get("gte")
+        lte = val_range.get("lte")
+        matching_values_pk = []
+
+        for value, pk in attr_values.items():
+            if is_date:
+                value = value.date()
+
+            if gte and lte:
+                if gte <= value <= lte:
+                    matching_values_pk.append(pk)
+            elif gte:
+                if gte <= value:
+                    matching_values_pk.append(pk)
+            elif lte >= value:
+                matching_values_pk.append(pk)
+
+        queries[attr_pk] += matching_values_pk
+
+
+def _clean_product_attributes_boolean_filter_input(filter_value, queries):
     attribute_slugs = [slug for slug, _ in filter_value]
     attributes = Attribute.objects.filter(
         input_type=AttributeInputType.BOOLEAN, slug__in=attribute_slugs
@@ -129,7 +164,7 @@ def _clean_product_attributes_boolean_filter_input(
         attr_pk = values_map[attr_slug]["pk"]
         value_pk = values_map[attr_slug]["values"].get(val)
         if value_pk:
-            queries[attr_pk] = [value_pk]
+            queries[attr_pk] += [value_pk]
 
 
 def filter_products_by_attributes_values(qs, queries: T_PRODUCT_FILTER_QUERIES):
@@ -168,7 +203,12 @@ def filter_products_by_attributes_values(qs, queries: T_PRODUCT_FILTER_QUERIES):
 
 
 def filter_products_by_attributes(
-    qs, filter_values, filter_range_values, filter_boolean_values
+    qs,
+    filter_values,
+    filter_range_values,
+    filter_boolean_values,
+    date_range_list,
+    date_time_range_list,
 ):
     queries: Dict[int, List[Optional[int]]] = defaultdict(list)
     try:
@@ -176,6 +216,14 @@ def filter_products_by_attributes(
             _clean_product_attributes_filter_input(filter_values, queries)
         if filter_range_values:
             _clean_product_attributes_range_filter_input(filter_range_values, queries)
+        if date_range_list:
+            _clean_product_attributes_date_time_range_filter_input(
+                date_range_list, queries, is_date=True
+            )
+        if date_time_range_list:
+            _clean_product_attributes_date_time_range_filter_input(
+                date_time_range_list, queries
+            )
         if filter_boolean_values:
             _clean_product_attributes_boolean_filter_input(
                 filter_boolean_values, queries
@@ -269,8 +317,10 @@ def filter_products_by_stock_availability(qs, stock_availability, channel_slug):
 def _filter_attributes(qs, _, value):
     if value:
         value_list = []
-        value_range_list = []
         boolean_list = []
+        value_range_list = []
+        date_range_list = []
+        date_time_range_list = []
 
         for v in value:
             slug = v["slug"]
@@ -278,10 +328,20 @@ def _filter_attributes(qs, _, value):
                 value_list.append((slug, v["values"]))
             elif "values_range" in v:
                 value_range_list.append((slug, v["values_range"]))
+            elif "date" in v:
+                date_range_list.append((slug, v["date"]))
+            elif "date_time" in v:
+                date_time_range_list.append((slug, v["date_time"]))
             elif "boolean" in v:
                 boolean_list.append((slug, v["boolean"]))
+
         qs = filter_products_by_attributes(
-            qs, value_list, value_range_list, boolean_list
+            qs,
+            value_list,
+            value_range_list,
+            boolean_list,
+            date_range_list,
+            date_time_range_list,
         )
     return qs